Crate tcp_port_scanner

Mutlithreaded TCP port scanner

Command Line Arguments

-h or --help - Prints the help menu

-j or --threads - Number of threads to use, max is 65535, default is 4

IpAddr - The IPv4 or IPv6 Address you want to scan in

Examples

./ip_sniffer 192.168.1.1 ./ip_sniffer -j 192.168.1.1

Notes

The command line arguments must come before the IP Address

Functions

scan

Function used to scan a port Returns a port number to the thread being used